packages are building and deploying, fixing client e2e test bugs

+++ packages/wasm/as/assembly/w3/subinvoke.ts
@@ -4,23 +4,23 @@
 
 @external("w3", "__w3_subinvoke")
 export declare function __w3_subinvoke(
-  uri_ptr: usize, uri_len: usize,
-  module_ptr: usize, module_len: usize,
-  method_ptr: usize, method_len: usize,
-  input_ptr: usize, input_len: usize
+  uri_ptr: u32, uri_len: u32,
+  module_ptr: u32, module_len: u32,
+  method_ptr: u32, method_len: u32,
+  input_ptr: u32, input_len: u32
 ): bool;
 
 // Subinvoke Result
 @external("w3", "__w3_subinvoke_result_len")
-export declare function __w3_subinvoke_result_len(): usize;
+export declare function __w3_subinvoke_result_len(): u32;
 @external("w3", "__w3_subinvoke_result")
-export declare function __w3_subinvoke_result(ptr: usize): void;
+export declare function __w3_subinvoke_result(ptr: u32): void;
 
 // Subinvoke Error
 @external("w3", "__w3_subinvoke_error_len")
-export declare function __w3_subinvoke_error_len(): usize;
+export declare function __w3_subinvoke_error_len(): u32;
 @external("w3", "__w3_subinvoke_error")
-export declare function __w3_subinvoke_error(ptr: usize): void;
+export declare function __w3_subinvoke_error(ptr: u32): void;
 
 // Subinvoke API Helper
 export function w3_subinvoke(
@@ -33,23 +33,23 @@ export function w3_subinvoke(
   const moduleBuf = String.UTF8.encode(module);
   const methodBuf = String.UTF8.encode(method);
   const success = __w3_subinvoke(
-    changetype<usize>(uriBuf), uriBuf.byteLength,
-    changetype<usize>(moduleBuf), moduleBuf.byteLength,
-    changetype<usize>(methodBuf), methodBuf.byteLength,
-    changetype<usize>(input), input.byteLength
+    changetype<u32>(uriBuf), uriBuf.byteLength,
+    changetype<u32>(moduleBuf), moduleBuf.byteLength,
+    changetype<u32>(methodBuf), methodBuf.byteLength,
+    changetype<u32>(input), input.byteLength
   );
 
   if (!success) {
     const errorLen = __w3_subinvoke_error_len();
-    const messageBuf = new ArrayBuffer(changetype<usize>(errorLen));
-    __w3_subinvoke_error(changetype<usize>(messageBuf));
+    const messageBuf = new ArrayBuffer(changetype<u32>(errorLen));
+    __w3_subinvoke_error(changetype<u32>(messageBuf));
     const message = String.UTF8.decode(messageBuf);
     throw new Error(message);
   }
 
   const resultLen = __w3_subinvoke_result_len();
-  const resultBuffer = new ArrayBuffer(changetype<usize>(resultLen));
-  __w3_subinvoke_result(changetype<usize>(resultBuffer));
+  const resultBuffer = new ArrayBuffer(changetype<u32>(resultLen));
+  __w3_subinvoke_result(changetype<u32>(resultBuffer));
 
   return resultBuffer;
 }
